V3.77 (#424)
* Incrementing SEMVER to v3.77

* enabling github Action for v3.77 branch

* redesigning the builders container setup into its own container base

* fixing educational examples with the new posit component API

* compilation fix for gcc

* dockerfile and build script maintenance

* enabling macos builds to RCA

* bug fix on release container build to tag latest correctly

* Dropping back to old MacOS-12 in CI to RCA breakage on 14

* WIP: making fixpnt isnan/isinf constexpr as Xcode 15 does as part of Cxx23

* WIP: removing isinf/isnan from fixpnt_fwd

* WIP: test: put isinf/isnan in the global namespace

* code hygiene and adding a complex playground to explore MacOS breakage

* enabling new build

* code hygiene: removing warnings blocksignificant

---------

+++ include/universal/native/ieee754_float.hpp
@@ -42,7 +42,7 @@ inline std::string to_hex(float number, bool nibbleMarker = false, bool hexPrefi
 	if (hexPrefix) s << "0x";
 	int nrNibbles = 8;
 	int nibbleIndex = (nrNibbles - 1);
-	uint32_t mask = (0xFull << (nibbleIndex * 4));
+	uint32_t mask = static_cast<uint32_t>(0xFull << (nibbleIndex * 4));
 	//	std::cout << "mask       : " << to_binary(mask, nbits) << '\n';
 	for (int n = nrNibbles - 1; n >= 0; --n) {
 		uint32_t raw = (bits & mask);
